Move pretty function to own file
